backend/requirements.txt
fastapi
uvicorn
pandas
pydantic
python-multipart
groq
python-dotenv
